feat(cli): add pipeline schema command
